+++ rosmon_core/src/monitor/node_monitor.h
@@ -143,7 +143,7 @@ public:
 	 *
 	 * @return Memory usage in bytes
 	 **/
-	inline double memory() const
+	inline uint64_t memory() const
 	{ return m_memory; }
 
 	inline unsigned int restartCount() const
@@ -152,7 +152,7 @@ public:
 	inline uint64_t memoryLimit() const
 	{ return m_launchNode->memoryLimitByte();}
 
-	inline float cpuLimit() const
+	inline double cpuLimit() const
 	{ return m_launchNode->cpuLimit();}
 
 	//@}
